Refactor generation of citation strings
